backend/app/models/uni_pipeline.py
"""
UNI 2 Pipeline
Harvard's pathology foundation model for tile-level feature extraction.
"""

import torch
import torch.nn as nn
import numpy as np
from typing import List, Optional
from pathlib import Path
import asyncio


class UNIPipeline:
    """
    Harvard UNI 2 foundation model for pathology tile encoding.

    Based on ViT-Large architecture trained on 100M+ pathology tiles.
    """

    def __init__(
        self,
        device: str = "cuda",
        model_version: str = "v2.0",
    ):
        self.device = device
        self.model_version = model_version
        self.model = None
        self.feature_dim = 1024  # ViT-Large hidden dim
        self.patch_size = 224

    async def load_model(self):
        """Load model weights from HuggingFace."""
        if self.model is not None:
            return

        # In production:
        # from transformers import AutoModel, AutoImageProcessor
        # self.model = AutoModel.from_pretrained(
        #     "MahmoodLab/UNI2-h",
        #     trust_remote_code=True,
        #     token=settings.huggingface_token
        # ).to(self.device)
        # self.processor = AutoImageProcessor.from_pretrained("MahmoodLab/UNI2-h")

        # Placeholder
        self.model = nn.Identity()
        print(f"Loaded UNI {self.model_version}")

    async def preprocess_tile(self, tile: np.ndarray) -> torch.Tensor:
        """
        Preprocess a tile for the model.

        Args:
            tile: RGB image array (H, W, C)

        Returns:
            Preprocessed tensor
        """
        # Standard ImageNet normalization for pathology
        mean = [0.485, 0.456, 0.406]
        std = [0.229, 0.224, 0.225]

        # Convert to tensor
        tensor = torch.from_numpy(tile).float().permute(2, 0, 1) / 255.0

        # Normalize
        for t, m, s in zip(tensor, mean, std):
            t.sub_(m).div_(s)

        return tensor

    async def extract_features(
        self,
        tiles: List[np.ndarray],
        batch_size: int = 64,
    ) -> torch.Tensor:
        """
        Extract features from a list of tiles.

        Args:
            tiles: List of tile images
            batch_size: Processing batch size

        Returns:
            Feature tensor of shape (num_tiles, feature_dim)
        """
        await self.load_model()

        all_features = []

        for i in range(0, len(tiles), batch_size):
            batch = tiles[i:i + batch_size]

            # Preprocess
            batch_tensors = []
            for tile in batch:
                tensor = await self.preprocess_tile(tile)
                batch_tensors.append(tensor)

            batch_tensor = torch.stack(batch_tensors).to(self.device)

            # Forward pass
            with torch.no_grad():
                # In production: features = self.model(batch_tensor).last_hidden_state[:, 0]
                features = torch.randn(batch_tensor.shape[0], self.feature_dim)

            all_features.append(features.cpu())

        return torch.cat(all_features, dim=0)

    async def extract_features_from_dir(
        self,
        tile_dir: str,
        batch_size: int = 64,
    ) -> torch.Tensor:
        """
        Extract features from tiles in a directory.

        Args:
            tile_dir: Path to directory containing tile images
            batch_size: Processing batch size

        Returns:
            Feature tensor
        """
        from PIL import Image

        tile_dir = Path(tile_dir)
        tile_files = sorted(tile_dir.glob("tile_*.png"))

        tiles = []
        for tile_file in tile_files:
            img = Image.open(tile_file).convert("RGB")
            tiles.append(np.array(img))

        return await self.extract_features(tiles, batch_size)

    async def get_embeddings(
        self,
        features: torch.Tensor,
    ) -> np.ndarray:
        """
        Get final embeddings for storage/retrieval.

        Args:
            features: Raw features from extract_features

        Returns:
            Normalized embeddings
        """
        # L2 normalize
        norms = torch.norm(features, p=2, dim=1, keepdim=True)
        normalized = features / (norms + 1e-8)

        return normalized.numpy()


# Utility function for integration
async def encode_slide_tiles(tile_dir: str, device: str = "cuda") -> np.ndarray:
    """
    Encode all tiles in a directory using UNI 2.

    Args:
        tile_dir: Directory containing tile images
        device: Device to use

    Returns:
        Numpy array of embeddings
    """
    pipeline = UNIPipeline(device=device)
    features = await pipeline.extract_features_from_dir(tile_dir)
    embeddings = await pipeline.get_embeddings(features)

    return embeddings