backend/requirements.txt
# FastAPI
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.2.1

# Database & Auth
supabase==2.3.4
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25

# Async & Tasks
celery==5.3.6
redis==5.0.1
httpx==0.26.0

# ML & Scientific
torch==2.2.0
torchvision==0.17.0
numpy==1.26.4
pillow==10.2.0
openslide-python==1.3.1
scikit-learn==1.4.0
pandas==2.2.0

# HuggingFace
transformers==4.37.2
huggingface-hub==0.20.3
safetensors==0.4.2

# Image processing
large-image==1.25.0
large-image-source-openslide==1.25.0
tifffile==2024.1.30

# PDF Generation
reportlab==4.0.8
weasyprint==60.1

# Monitoring
sentry-sdk[fastapi]==1.40.0

# Stripe
stripe==8.4.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0